AltayAnimal > MCAnimal && created new class for the pm animal
