+++ src/pocketmine/entity/behavior/MateBehavior.php
@@ -24,7 +24,7 @@ declare(strict_types=1);
 
 namespace pocketmine\entity\behavior;
 
-use pocketmine\entity\Animal;
+use pocketmine\entity\MCAnimal;
 use pocketmine\entity\Entity;
 use pocketmine\entity\Mob;
 
@@ -33,10 +33,10 @@ class MateBehavior extends Behavior{
 	protected $speedMultiplier;
 	/** @var int */
 	protected $spawnBabyDelay = 0;
-	/** @var Animal */
+	/** @var MCAnimal */
 	protected $targetMate;
 
-	public function __construct(Animal $mob, float $speedMultiplier){
+	public function __construct(MCAnimal $mob, float $speedMultiplier){
 		parent::__construct($mob);
 
 		$this->speedMultiplier = $speedMultiplier;
@@ -75,13 +75,13 @@ class MateBehavior extends Behavior{
 		$this->spawnBabyDelay = 0;
 	}
 
-	public function getNearbyMate() : ?Animal{
+	public function getNearbyMate() : ?MCAnimal{
 		$list = $this->mob->level->getEntities();
 		$dist = 8;
 		$animal = null;
 
 		foreach($list as $entity){
-			if($entity !== $this->mob and $entity instanceof Animal and $entity->isInLove() and !$entity->isBaby() and $entity->distance($this->mob) < $dist and $entity::NETWORK_ID === $this->mob::NETWORK_ID){
+			if($entity !== $this->mob and $entity instanceof MCAnimal and $entity->isInLove() and !$entity->isBaby() and $entity->distance($this->mob) < $dist and $entity::NETWORK_ID === $this->mob::NETWORK_ID){
 				$dist = $entity->distance($this->mob);
 				$animal = $entity;
 			}
